added new valueObjects for Special Days
