+++ osgp-domain-core/src/main/java/com/alliander/osgp/domain/core/valueobjects/smartmetering/SpecialDay.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2015 Smart Society Services B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ */
+package com.alliander.osgp.domain.core.valueobjects.smartmetering;
+
+import java.io.Serializable;
+
+public class SpecialDay implements Serializable {
+
+    private static final long serialVersionUID = 8009905447322732462L;
+
+    protected String specialDayDate;
+    
+    protected int dayId;
+
+    public String getSpecialDayDate() {
+        return specialDayDate;
+    }
+
+    public void setSpecialDayDate(String specialDayDate) {
+        this.specialDayDate = specialDayDate;
+    }
+
+    public int getDayId() {
+        return dayId;
+    }
+
+    public void setDayId(int dayId) {
+        this.dayId = dayId;
+    }
+}
